:art: TaskManager logic finished :white_check_mark:
